+++ data/src/buffer.rs
@@ -146,12 +146,29 @@ impl From<config::Buffer> for Settings {
     }
 }
 
+#[derive(Debug, Clone, Deserialize)]
+#[serde(default)]
+pub struct TextInputNickname {
+    pub enabled: bool,
+    pub show_access_level: bool,
+}
+
+impl Default for TextInputNickname {
+    fn default() -> Self {
+        Self {
+            enabled: true,
+            show_access_level: true,
+        }
+    }
+}
+
 #[derive(Debug, Clone, Default, Deserialize)]
 #[serde(default)]
 pub struct TextInput {
     pub visibility: TextInputVisibility,
     pub auto_format: AutoFormat,
     pub autocomplete: Autocomplete,
+    pub nickname: TextInputNickname,
 }
 
 #[derive(Debug, Clone, Copy, Default, Deserialize)]