+++ src/screen/dashboard/pane.rs
@@ -410,7 +410,7 @@ impl TitleBar {
                 .shaping(text::Shaping::Advanced),
         )
         .height(theme::line_height(&config.font).ceil().max(22.0))
-        .padding([0, 10])
+        .padding([0, 4])
         .align_y(iced::alignment::Vertical::Center);
 
         widget::TitleBar::new(title)